[3870] [Time Beats: 100.00%] [Memory Beats: 15.44%] - Leet2Hub

+++ 13-Math/3870. Count Commas in Range/3870. Count Commas in Range.cpp
@@ -1,7 +1,7 @@
 1class Solution {
 2public:
 3    int countCommas(int n) {
-4        long long ans = 0;
+4        int ans = 0;
 5        for (long long x = 1000; x <= n; x *= 1000) {
 6            ans += n - x + 1;
 7        }
